api/requirements/core.txt
# Core Layer Dependencies (~5MB)
# Used by ALL Lambda functions

boto3>=1.34.0
botocore>=1.34.0
python-jose[cryptography]>=3.3.0
pydantic>=2.5.0
requests>=2.31.0